requirements.txt
streamlit>=1.30.0
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
vaderSentiment>=3.3.2
feedparser>=6.0.0
requests>=2.31.0
scikit-learn>=1.3.0
scipy>=1.11.0
plotly>=5.18.0
transformers>=4.36.0
torch>=2.1.0
python-dotenv>=1.0.0
